+++ apps/api/src/app/portfolio/current-rate.service.ts
@@ -1,4 +1,4 @@
-import { OrderService } from '@ghostfolio/api/app/order/order.service';
+import { ActivitiesService } from '@ghostfolio/api/app/activities/activities.service';
 import { LogPerformance } from '@ghostfolio/api/interceptors/performance-logging/performance-logging.interceptor';
 import { DataProviderService } from '@ghostfolio/api/services/data-provider/data-provider.service';
 import { MarketDataService } from '@ghostfolio/api/services/market-data/market-data.service';
@@ -24,9 +24,9 @@ export class CurrentRateService {
   private static readonly MARKET_DATA_PAGE_SIZE = 50000;
 
   public constructor(
+    private readonly activitiesService: ActivitiesService,
     private readonly dataProviderService: DataProviderService,
     private readonly marketDataService: MarketDataService,
-    private readonly orderService: OrderService,
     @Inject(REQUEST) private readonly request: RequestWithUser
   ) {}
 
@@ -129,10 +129,11 @@ export class CurrentRateService {
 
           if (!value) {
             // Fallback to unit price of latest activity
-            const latestActivity = await this.orderService.getLatestOrder({
-              dataSource,
-              symbol
-            });
+            const latestActivity =
+              await this.activitiesService.getLatestActivity({
+                dataSource,
+                symbol
+              });
 
             value = {
               dataSource,