+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Included asset profile data in the holdings of the public page
 - Reused the value component in the platform management of the admin control panel
 - Reused the value component in the tag management of the admin control panel
+- Deprecated the `api/v1/order` endpoints in favor of the `api/v1/activities` endpoints
 - Upgraded `jsonpath` from version `1.1.1` to `1.2.1`
 
 ### Fixed